% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{educational}
\alias{educational}
\title{Observational Educational Dataset}
\format{
A data frame with 1000 rows and 7 variables:
\describe{
\item{SAT_MATH}{SAT Math Score}
\item{HSGPA}{High School GPA}
\item{AGE}{Age of Student}
\item{GENDER}{Gender of Student}
\item{URM}{Under Represented Minority}
\item{A}{Treatment Variable}
\item{Y}{Students Final Grade}
}
}
\source{
Wilke, Morten C., et al. “Estimating the Optimal Treatment
Regime for Student Success Programs.” Behaviormetrika, vol. 48, no. 2, 2021,
pp. 309–343., https://doi.org/10.1007/s41237-021-00140-0.
}
\usage{
educational
}
\description{
A simulated dataset containing the grades and other attributes of 1000
simulated students
}
\keyword{datasets}
