% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_split}
\alias{find_split}
\title{Multiply a number by two}
\usage{
find_split(y, x, trt, cutpts, method, propensity, minbucket, response_type)
}
\arguments{
\item{x}{A single integer.}
}
\description{
Multiply a number by two
}
