% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.CERFIT}
\alias{predict.CERFIT}
\title{Get predictions from a CERFIT object}
\usage{
\method{predict}{CERFIT}(
  object,
  data,
  newdata,
  gridval = NULL,
  prediction = c("overall", "by iter"),
  type = c("response", "ITE", "node", "opT"),
  alpha = 0.5,
  useRse = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class CERFIT}

\item{data}{Sata used to build the tree}

\item{newdata}{New data to make predictions from}

\item{gridval}{For continuous treatment. Controls for what values of treatment to predict}

\item{prediction}{Return prediction using all trees ("overall") or using first i trees ("by iter")}

\item{type}{Choose what you want ro predict}

\item{alpha}{something}

\item{useRse}{THis will be removed}

\item{...}{Additional Arguments}
}
\description{
Get predictions from a CERFIT object
}
