% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{warts}
\alias{warts}
\title{Randomized Controlled Trial Warts Dataset}
\format{
A data frame with 180 rows and 8 variables:
\describe{
\item{sex}{Patients Sex}
\item{age}{Patients Age}
\item{Time}{Time Elapsed Before Treatment}
\item{Number_of_Warts}{Number of Warts}
\item{Type}{Type of Wart}
\item{Area}{Wart Surface Area}
\item{Result_of_Treatment}{Treatment Outcome}
\item{treatment}{0 for immunotherapy and 1 for cryotherapy}
}
}
\source{
Khozeimeh, Fahime, et al. “An Expert System for Selecting Wart
Treatment Method.” Computers in Biology and Medicine, vol. 81, 2017,
pp. 167–175., https://doi.org/10.1016/j.compbiomed.2017.01.001.
}
\usage{
warts
}
\description{
A dataset comparing immunotherapy to cryotherapy treatments and their effeteness of
removing warts
}
\keyword{datasets}
