% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CERFIT.R
\name{CERFIT}
\alias{CERFIT}
\title{Fits a Causal Effect Random Forest of Interactions Trees}
\usage{
CERFIT(
  formula,
  data,
  ntrees,
  subset = NULL,
  search = c("exhaustive", "sss"),
  method = c("RCT", "observation"),
  PropForm = c("randomForest", "CBPS", "GBM", "HI"),
  split = c("t.test"),
  mtry = NULL,
  nsplit = NULL,
  nsplit.random = TRUE,
  minsplit = 20,
  minbucket = round(minsplit/3),
  maxdepth = 30,
  a = 50,
  sampleMethod = c("bootstrap", "subsample", "subsampleByID", "learning"),
  useRes = TRUE,
  scale.y = FALSE
)
}
\arguments{
\item{formula}{Formula to build CERFIT.  Categorical predictors must be listed as a factor. e.g., Y ~ x1 + x2 | treatment}

\item{data}{Data to grow a tree.}

\item{ntrees}{Number of Trees to grow}

\item{subset}{thing}

\item{search}{Method to search through candidate splits}

\item{method}{For observational study data, method="observation";for randomized study data, method="RCT".}

\item{PropForm}{Method to estimate propensity score}

\item{split}{Impurity measure splitting statistic}

\item{mtry}{Number of variables to consider at each split}

\item{nsplit}{Number of cut points selected}

\item{nsplit.random}{Logical: indicates if process to select cut points are random}

\item{minsplit}{Number of observations required to continue growing tree}

\item{minbucket}{Number of observations required in each child node}

\item{maxdepth}{Maximum depth of tree}

\item{a}{Sigmoid approximation variable (for "sss" which is still under development)}

\item{sampleMethod}{Method to sample learning sample}

\item{useRes}{Logical indicator if you want to fit the CERFIT model to linear model}

\item{scale.y}{Logical, standardize y when creating splits (For "sss" to increase stability)}
}
\value{
Returns a fitted CERFIT object which is a list with the following elements
\itemize{
\item RandFor: The Random forest of interaction trees
\item trt.type: A string containing the treatment type of the data used to fit the model
\item response.type: A string representing the response type of the data
\item useRes: A logical indicator that is TRUE if the model was fit on the
resdiuals of a linear model
\item data: The data used to fit the model also contains the propensity score if
method was set to observational}
}
\description{
Fits a Causal Effect Random Forest of Interactions Trees
}
\details{
This function is implementation of Random Forest of Interaction Trees. A
decision tree based method that approximates the effect of a treatment on each individual in
a dataset. It does this by estimating the Individualized Treatment Effect (ITE) for each
observation. This function also implements the extension of RFIT to observational data and
to multiple and ordered treatments.
}
