% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CERFIT.R
\name{CERFIT}
\alias{CERFIT}
\title{Fits a Random Forest of Interactions Trees}
\usage{
CERFIT(
  formula,
  data,
  ntrees,
  subset = NULL,
  search = c("exhaustive", "sss"),
  method = c("RCT", "observation"),
  PropForm = c("randomForest", "CBPS", "GBM", "HI"),
  split = c("t.test"),
  mtry = NULL,
  nsplit = NULL,
  nsplit.random = TRUE,
  minsplit = 20,
  minbucket = round(minsplit/3),
  maxdepth = 30,
  a = 50,
  sampleMethod = c("bootstrap", "subsample", "subsampleByID", "learning"),
  useRes = TRUE,
  scale.y = FALSE
)
}
\arguments{
\item{formula}{Formula to build CERFIT.  Categorical predictors must be listed as a factor. e.g., Y ~ x1 + x2 | treatment}

\item{data}{Data to grow a tree.}

\item{ntrees}{Number of Trees to grow}

\item{subset}{thing}

\item{search}{Method to search through candidate splits}

\item{method}{For observational study data, method="observation";for randomized study data, method="RCT".}

\item{PropForm}{Method to estimate propensity score}

\item{split}{Impurity measure splitting statistic}

\item{mtry}{Number of variables to consider at each split}

\item{nsplit}{Number of cut points selected}

\item{nsplit.random}{Logical: indicates if process to select cut points are random}

\item{minsplit}{Number of observations required to continue growing tree}

\item{minbucket}{Number of observations required in each child node}

\item{maxdepth}{Maximum depth of tree}

\item{a}{Sigmoid approximation variable (for "sss" which is still under development)}

\item{sampleMethod}{Method to sample learning sample}

\item{useRes}{Logical indicator if you want to fit the CERFIT model to linear model}

\item{scale.y}{Logical, standardize y when creating splits (For "sss" to increase stability)}
}
\value{
Returns a fitted CERFIT object which is a list with the following elements
\itemize{
\item RandFor: The Random forest of interaction trees
\item trt.type: A string containing the treatment type of the data used to fit the model
\item response.type: A string representing the response type of the data
\item useRes: A logical indicator that is TRUE if the model was fit on the
residuals of a linear model
\item data: The data used to fit the model also contains the propensity score if
method was set to observational}
}
\description{
Estimates the a observations response for each level of treatment
}
\details{
This function is implementation of Random Forest of Interaction Trees proposed
in Su (2018). It also handles extension for multiple, ordered and continuous
treatment. Function can be used for RCT data or observational data has shown in Li, et al.
(2022). The RFIT is decision tree based method that approximates the effect of a treatment
on each individual in a dataset. It does this by estimating a observations response
for each level of treatment.
}
\references{
\itemize{
\item Li, Luo, et al. Causal Effect Random Forest of
Interaction Trees for Learning Individualized Treatment Regimes with
Multiple Treatments in Observational Studies. Stat, 2022,
https://doi.org/10.1002/sta4.457.
\item Su, X., Peña, A., Liu, L., & Levine, R. (2018). Random forests of interaction trees for estimating individualized treatment effects in randomized trials.
Statistics in Medicine, 37(17), 2547- 2560.}
}
