% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CERFIT.R
\name{CERFIT}
\alias{CERFIT}
\title{Fits a Causal Effect Random Forest of Interactions Trees}
\usage{
CERFIT(
  formula,
  data,
  ntrees,
  subset = NULL,
  search = c("exhaustive", "sss"),
  method = c("RCT", "observation"),
  PropForm = c("randomForest", "CBPS", "GBM", "HI"),
  split = c("t.test"),
  mtry = NULL,
  nsplit = NULL,
  nsplit.random = TRUE,
  minsplit = 20,
  minbucket = round(minsplit/3),
  maxdepth = 30,
  a = 50,
  sampleMethod = c("bootstrap", "subsample", "subsampleByID", "learning"),
  useRes = TRUE,
  scale.y = FALSE
)
}
\arguments{
\item{formula}{Formula to build CERFIT.  Categorical predictors must be listed as a factor. e.g., Y ~ x1 + x2 | treatment}

\item{data}{Data to grwo a tree.}

\item{search}{Method to search throught candidate splits}

\item{method}{For observational stuy data, method="observation";for randomized study data, method="RCT".}

\item{PropForm}{Method to estimate propensity score}

\item{split}{Impurity measure splitting statistic}

\item{mtry}{Number of varialbes to consider at each split}

\item{nsplit}{Number of cut points selected}

\item{nsplit.random}{Logical: indicates if process to select cutpts are random}

\item{minsplit}{Number of observations required to continue growing tree}

\item{minbucket}{Number of observations required in each child node}

\item{maxdepth}{Maximum depth of tree}

\item{a}{Sigmoid approximation variable (for "sss" which is still under development)}

\item{sampleMethod}{Method to sample learning sample}

\item{scale.y}{Logical, standardize y when creating splits (For "sss" to increase stability)}
}
\value{
The fitted CERFIT model.
}
\description{
Fits a Causal Effect Random Forest of Interactions Trees
}
