% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.CERFIT}
\alias{predict.CERFIT}
\title{Get predictions from a CERFIT object}
\usage{
\method{predict}{CERFIT}(
  object,
  newdata = object$data,
  gridval = NULL,
  prediction = c("overall", "by iter"),
  type = c("response", "ITE", "node", "opT"),
  alpha = 0.5,
  ...
)
}
\arguments{
\item{object}{An object of class CERFIT}

\item{newdata}{New data to make predictions from}

\item{gridval}{For continuous treatment. Controls for what values of treatment to predict}

\item{prediction}{Return prediction using all trees ("overall") or using first i trees ("by iter")}

\item{type}{Choose what value you wish to predict. Response will predict the repsone.
ITE will predict the Individialzed treatment effect. Node will predcit the node. And opT
will predict the optimal treatment for each observation.}

\item{alpha}{something}

\item{...}{Additional Arguments}
}
\value{
A named vector where the values are the average minimum depth for each variable
and the names are the associated variable names
}
\description{
Get predictions from a CERFIT object
}
