% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.CERFIT}
\alias{predict.CERFIT}
\title{Get predictions from a CERFIT object}
\usage{
\method{predict}{CERFIT}(
  object,
  newdata = NULL,
  gridval = NULL,
  prediction = c("overall", "by iter"),
  type = c("response", "ITE", "node", "opT"),
  alpha = 0.5,
  ...
)
}
\arguments{
\item{object}{A fitted CERFIT object}

\item{newdata}{New data to make predictions from. IF not provided will make predctions
on training data}

\item{gridval}{For continuous treatment. Controls for what values of treatment to predict}

\item{prediction}{Return prediction using all trees ("overall") or using first i trees ("by iter")}

\item{type}{Choose what value you wish to predict. Response will predict the response.
ITE will predict the Individualized treatment effect. Node will predict the node. And opT
will predict the optimal treatment for each observation.}

\item{alpha}{For continuous treatment its is the mixing parameter for the elastic
net regularization in each node. When equal to 0 it is ridge regression and
when equal to 1 is is lasso regression.}

\item{...}{Additional Arguments}
}
\value{
The return value depends of the type argument. If type is response the function
will return matrix a n rows and a number of columns equal to the levels of treatment.
IF type is ITE then it returns a matrix with n rows and a number of columns equal to
one minus the levels of treatment. And if type is opT then its return a numeric
vector of length n.
}
\description{
Get predictions from a CERFIT object
}
\examples{
data <- data.frame(y = rnorm(100),x = rnorm(100),t = rbinom(1000,1,.5))
fit <- CERFIT(y ~ x | t,method = "RCT",data = data,ntrees = 10)
ite <- predict(fit,type = "ITE")
}
