% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.CERFIT}
\alias{predict.CERFIT}
\title{Get predictions from a CERFIT object}
\usage{
\method{predict}{CERFIT}(
  x,
  data,
  newdata,
  gridval = NULL,
  prediction = c("overall", "by iter"),
  type = c("response", "ITE", "node", "opT"),
  alpha = 0.5,
  useRse = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class CERFIT}

\item{data}{data used to build the tree}

\item{newdata}{new data to make predictions from}

\item{gridval}{something that exists}

\item{prediction}{Return prediction using all trees ("overall") or using first i trees ("by iter")}

\item{type}{Choose what you want ro predict}

\item{alpha}{something}

\item{useRse}{something}

\item{...}{something}
}
\description{
Get predictions from a CERFIT object
}
